+++ login.html
@@ -4,12 +4,13 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>login</title>
+    <script src="auth.js"></script>
 </head>
 <body>
-    <input type="email">
+    <input type="email" id="email">
     <br>
-    <input type="password">
+    <input type="password" id="password">
     <br>
-    <button>login</button>
+    <button id="login">login</button>
 </body>
 </html>
